x86/resctrl: Make resctrl_arch_rmid_read() safe to call in IRQ context
Resctrl always calls resctrl_arch_rmid_read() in process context with
IRQs unmasked. This is because MPAM needs IPI remote CPUs that are part
of the same domain, (when additional hardware structures need to be read
to read the domain counter), and to sleep if the counter takes time to
settle.

The conflicts with the behaviour needed for a PMU driver, where
resctrl_arch_rmid_read() might be called in both process and irq context.
MPAM systems that require the IPI are unable to use the PMU driver.

For irq context it is safe to assume resctrl_arch_rmid_read() is called
on one of the CPUs in the domain. If not, it can fail.

Change x86's resctrl_arch_rmid_read() to skip smp_call_function_any()
if interrupts are masked and it's called on a CPU in the correct domain.

Signed-off-by: James Morse <james.morse@arm.com>