arm_mpam: resctrl: detect priority partitioning capability
ARM MPAM supports different control that can be appiled to different
resources in the system, for instace priority partitioning control
that can be appiled over system interconnects, and propagated from
one MSC to other (known as downstream priority), or can be appiled
within an MSC for internal operations.

This change lets resctrl know the about, if particular resource has
the priority partitioning capability.

Signed-off-by: Amit Singh Tomar <amitsinght@marvell.com>